refactor(bench): share connection setup and schedule
