+++ ydb/benches/sdk_compare/topic.rs
@@ -1,20 +1,19 @@
 mod reader;
 mod writer;
 
-use std::env;
 use std::time::{Duration, Instant};
 
 use anyhow::{Context, Result};
 use tokio::time::timeout_at;
-use ydb::{Client, ClientBuilder, Codec, ConsumerBuilder, CreateTopicOptionsBuilder, TopicClient};
+use ydb::{Codec, ConsumerBuilder, CreateTopicOptionsBuilder, TopicClient};
 
 use crate::config::{Scenario, TopicWorkload};
+use crate::connection;
 use crate::result::{BenchmarkResult, TopicMetrics};
-
-const CONNECTION_STRING_ENV: &str = "YDB_CONNECTION_STRING";
+use crate::schedule::BenchmarkSchedule;
 
 pub(crate) async fn run(scenario: &Scenario, workload: &TopicWorkload) -> Result<BenchmarkResult> {
-    let client = connect().await?;
+    let client = connection::connect().call().await?;
     let topic_path = format!(
         "{}/{}",
         client.database().trim_end_matches('/'),
@@ -28,23 +27,6 @@ pub(crate) async fn run(scenario: &Scenario, workload: &TopicWorkload) -> Result
         eprintln!("warning: failed to drop benchmark topic {topic_path}: {error}");
     }
     result
-}
-
-async fn connect() -> Result<Client> {
-    let connection_string = env::var(CONNECTION_STRING_ENV)
-        .with_context(|| format!("{CONNECTION_STRING_ENV} is not set"))?;
-
-    let client = ClientBuilder::new_from_connection_string(&connection_string)
-        .context("failed to parse YDB connection string")?
-        .client()
-        .context("failed to create YDB client")?;
-
-    client
-        .wait()
-        .await
-        .context("failed to initialize YDB client")?;
-
-    Ok(client)
 }
 
 async fn create_topic(
@@ -124,57 +106,7 @@ async fn run_workload(
 }
 
 #[derive(Clone, Copy)]
-pub(super) struct BenchmarkSchedule {
-    origin: Instant,
-    pub(super) measurement_start: Instant,
-    pub(super) measurement_end: Instant,
-    pub(super) completion_deadline: Instant,
-}
-
-#[derive(Clone, Copy)]
 pub(super) enum AckLatencyStart {
     Warmup,
     Measured(Instant),
-}
-
-impl BenchmarkSchedule {
-    fn new(
-        warmup_duration: Duration,
-        measurement_duration: Duration,
-        drain_timeout: Duration,
-    ) -> Result<Self> {
-        let origin = Instant::now();
-        let measurement_start = origin
-            .checked_add(warmup_duration)
-            .context("warm-up deadline overflowed")?;
-        let measurement_end = measurement_start
-            .checked_add(measurement_duration)
-            .context("measurement deadline overflowed")?;
-        let completion_deadline = measurement_end
-            .checked_add(drain_timeout)
-            .context("measurement drain deadline overflowed")?;
-
-        Ok(Self {
-            origin,
-            measurement_start,
-            measurement_end,
-            completion_deadline,
-        })
-    }
-
-    pub(super) fn is_measurement_instant(&self, instant: Instant) -> bool {
-        instant >= self.measurement_start && instant < self.measurement_end
-    }
-
-    pub(super) fn ns_at(&self, instant: Instant) -> Result<u64> {
-        let elapsed = instant
-            .checked_duration_since(self.origin)
-            .context("benchmark instant is before schedule origin")?;
-        u64::try_from(elapsed.as_nanos())
-            .context("benchmark schedule does not fit into u64 nanoseconds")
-    }
-
-    pub(super) fn now_ns(&self) -> Result<u64> {
-        self.ns_at(Instant::now())
-    }
 }
